requirements.txt
fastapi
uvicorn
pypdf2
sentence-transformers
chromadb
google-generativeai
python-multipart
python-dotenv
tiktoken
